Revert "TDD, Port, Kalman, Telemetry, and Cleanup"
