+++ prototype/bezier.h
@@ -8,7 +8,7 @@
 #define _BEZIER_H_
 #include "common.h"
 
-void initMap(float* map);
+void initMap(void);
 /* Initialize the control points defining the map. */
 
 vector getDesiredHeading(vector currentPos);
