+++ prototype/motion.h
@@ -1,0 +1,14 @@
+/*
+ * Sierra College Computer Engineering Club: Sparkfun AVC
+ * Motion Subsystem
+ * Authors:
+ */
+
+#ifndef _MOTION_H_
+#define _MOTION_H_
+
+void initMotion();
+void updateMotion();
+
+#endif
+
