+++ prototype/nav.h
@@ -7,6 +7,12 @@
 #ifndef _NAV_H_
 #define _NAV_H_
 
+#define M1HEADING 280
+
+/*
+  I'm assuming true north for this milestone.
+*/
+
 typedef struct s_motionData
 {
 	/* In degrees */
@@ -18,7 +24,6 @@ typedef struct s_motionData
 void initNav();
 void updateNav();
 motionData* getMotionData();
-
 
 #endif
