+++ src/inc/navigation.h
@@ -1,0 +1,13 @@
+/*
+ * Sierra College Computer Engineering Club: Sparkfun AVC
+ * Navigation Subsystem
+ * authors: Kris Semelka, Richard Prasad
+ */
+
+#ifndef _SPARKFUN_NAVIGATION_H_
+#define _SPARKFUN_NAVIGATION_H_
+
+void navigationInit();
+void navigationUpdate();
+
+#endif
