+++ src/navigation.c
@@ -1,0 +1,14 @@
+/*
+ * Sierra College Computer Engineering Club: Sparkfun AVC
+ * Navigation Subsystem
+ * authors: Kris Semelka, Richard Prasad
+ */
+#include <stdio.h>
+
+void navigationInit()
+{
+}
+
+void navigationUpdate()
+{
+}
